Move server certificate to server module
