+++ feature/account/server/certificate/src/main/kotlin/app/k9mail/feature/account/server/certificate/domain/ServerCertificateDomainContract.kt
@@ -1,0 +1,21 @@
+package app.k9mail.feature.account.server.certificate.domain
+
+import app.k9mail.feature.account.server.certificate.domain.entity.ServerCertificateError
+import java.security.cert.X509Certificate
+
+interface ServerCertificateDomainContract {
+
+    interface ServerCertificateErrorRepository {
+        fun getCertificateError(): ServerCertificateError?
+
+        fun setCertificateError(serverCertificateError: ServerCertificateError)
+
+        fun clearCertificateError()
+    }
+
+    interface UseCase {
+        fun interface AddServerCertificateException {
+            suspend fun addCertificate(hostname: String, port: Int, certificate: X509Certificate?)
+        }
+    }
+}
