+++ feature/account/server/certificate/src/main/kotlin/app/k9mail/feature/account/server/certificate/domain/entity/ServerCertificateError.kt
@@ -1,0 +1,9 @@
+package app.k9mail.feature.account.server.certificate.domain.entity
+
+import java.security.cert.X509Certificate
+
+data class ServerCertificateError(
+    val hostname: String,
+    val port: Int,
+    val certificateChain: List<X509Certificate>,
+)
