+++ feature/account/server/validation/build.gradle.kts
@@ -27,7 +27,7 @@ dependencies {
 
     implementation(projects.feature.account.common)
     implementation(projects.feature.account.oauth)
-    implementation(projects.feature.account.servercertificate)
+    implementation(projects.feature.account.server.certificate)
 
     testImplementation(projects.core.ui.compose.testing)
 }
