--- feature/account/servercertificate/src/main/kotlin/app/k9mail/feature/account/servercertificate/domain/entity/ServerCertificateError.kt
@@ -1,9 +1,0 @@
-package app.k9mail.feature.account.servercertificate.domain.entity
-
-import java.security.cert.X509Certificate
-
-data class ServerCertificateError(
-    val hostname: String,
-    val port: Int,
-    val certificateChain: List<X509Certificate>,
-)
