+++ feature/account/setup/build.gradle.kts
@@ -29,7 +29,7 @@ dependencies {
 
     api(projects.feature.account.common)
     implementation(projects.feature.account.oauth)
-    implementation(projects.feature.account.servercertificate)
+    implementation(projects.feature.account.server.certificate)
     api(projects.feature.account.server.validation)
 
     testImplementation(projects.core.ui.compose.testing)
