+++ feature/account/setup/src/test/kotlin/app/k9mail/feature/account/setup/AccountSetupModuleKtTest.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import app.k9mail.core.common.oauth.OAuthConfigurationFactory
 import app.k9mail.feature.account.common.domain.entity.AccountState
 import app.k9mail.feature.account.oauth.ui.AccountOAuthContract
+import app.k9mail.feature.account.server.certificate.ui.ServerCertificateErrorContract
 import app.k9mail.feature.account.server.validation.ui.ServerValidationContract
-import app.k9mail.feature.account.servercertificate.ui.AccountServerCertificateErrorContract
 import app.k9mail.feature.account.setup.AccountSetupExternalContract.AccountCreator
 import app.k9mail.feature.account.setup.AccountSetupExternalContract.AccountCreator.AccountCreatorResult
 import app.k9mail.feature.account.setup.ui.AccountSetupContract
@@ -69,7 +69,7 @@ class AccountSetupModuleKtTest : KoinTest {
                 AccountOutgoingConfigContract.State::class,
                 AccountOptionsContract.State::class,
                 AccountState::class,
-                AccountServerCertificateErrorContract.State::class,
+                ServerCertificateErrorContract.State::class,
                 AuthStateStorage::class,
                 Context::class,
                 Boolean::class,
