+++ settings.gradle.kts
@@ -46,8 +46,8 @@ include(
 include(
     ":feature:account:common",
     ":feature:account:oauth",
-    ":feature:account:servercertificate",
     ":feature:account:setup",
+    ":feature:account:server:certificate",
     ":feature:account:server:validation",
 )
